HSEARCH-3975 Reorganize tests related to Hibernate ORM adapters
